GitOfHolding: Added support to backup game options
Summary: In this commit a minor issue was fixed with overwriting
savegames if the actual folder is empty. Also added support for backing
up the options file ("Baldur.lua" in the case of the supported games as
of today) as they seem to contain all the options set in game.

Test Plan: Downloaded the other games, "Baldur's Gate: Enhanced Edition"
and "Icewind Dale: Enhanced Edition" and only launched game and tinkered
with the options but not started a actual session to create a savegame
yet. Run script multiple times to check for correct results.